feat: beta option 추가
